+++ src/main/java/roomescape/service/ThemeService.java
@@ -2,8 +2,14 @@ package roomescape.service;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.common.config.ClockProvider;
+import roomescape.common.exception.AlreadyExistException;
+import roomescape.common.exception.NotFoundException;
+import roomescape.common.exception.UnprocessableException;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.Theme;
 import roomescape.dto.request.ThemeRequest;
@@ -15,22 +21,19 @@ public class ThemeService {
     private static final int POPULAR_THEME_PERIOD_DAYS = 6;
 
     private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
+    private final ClockProvider clockProvider;
 
-    public ThemeService(ThemeDao themeDao) {
+    public ThemeService(ThemeDao themeDao, ReservationDao reservationDao, ClockProvider clockProvider) {
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
+        this.clockProvider = clockProvider;
     }
 
     public ThemeResponse addTheme(ThemeRequest request) {
         validateUniqueTheme(request.name());
         Theme savedTheme = themeDao.insert(request.toTheme());
         return ThemeResponse.from(savedTheme);
-    }
-
-    private void validateUniqueTheme(String name) {
-        boolean exists = themeDao.existsByName(name);
-        if (exists) {
-            throw new IllegalArgumentException("이미 존재하는 테마입니다.");
-        }
     }
 
     @Transactional(readOnly = true)
@@ -41,18 +44,37 @@ public class ThemeService {
     }
 
     @Transactional(readOnly = true)
-    public List<ThemeResponse> getPopularThemes(LocalDate endDate) {
+    public List<ThemeResponse> getPopularThemes() {
+        LocalDate endDate = LocalDate.now(clockProvider.getClock());
         LocalDate startDate = endDate.minusDays(POPULAR_THEME_PERIOD_DAYS);
+
         List<Theme> popularThemes = themeDao.selectPopularThemesByPeriod(startDate, endDate.minusDays(1));
         return popularThemes.stream()
                 .map(ThemeResponse::from)
                 .toList();
     }
 
-    public void delete(long themeId) {
-        int deleted = themeDao.delete(themeId);
-        if (deleted == 0) {
-            throw new IllegalArgumentException("존재하지 않는 테마입니다.");
+    public void deleteTheme(long themeId) {
+        Optional<Theme> theme = themeDao.selectById(themeId);
+        if (theme.isEmpty()) {
+            throw new NotFoundException("존재하지 않는 테마입니다.");
+        }
+
+        validateThemeIncludeReservation(themeId);
+        themeDao.delete(themeId);
+    }
+
+    private void validateUniqueTheme(String name) {
+        boolean exists = themeDao.existsByName(name);
+        if (exists) {
+            throw new AlreadyExistException("이미 존재하는 테마입니다.");
+        }
+    }
+
+    private void validateThemeIncludeReservation(long themeId) {
+        boolean existsByThemeId = reservationDao.existsByThemeId(themeId);
+        if (existsByThemeId) {
+            throw new UnprocessableException("예약된 테마는 삭제할 수 없습니다.");
         }
     }
 }
